data-augmentation/requirements.txt
tensorflow
python-dotenv
h5py
opencv-python
matplotlib